Use listener script before injecting main hook
This has basically the same reasoning as the reduce async commit
earlier. This way, we know for a fact that Webpack's chunks have loaded
by the time our script has been loaded, so searching through and finding
packages is much less painful. This also requires an extra build step
because the payloads must be compiled to JavaScript so the injector can
eval them in the window context.